requirements.txt
Flask
google-generativeai
firebase-admin
qrcode[pil]
